*file size is now long in ES mapping (needed when file size exceeds 32-bit boundary)

+++ cmsDataFlowMerger.py
@@ -89,7 +89,7 @@ def esMonitorMapping(esServerUrl,esIndexName,numberOfShards,numberOfReplicas,deb
                'processed'     :{'type':'integer'},
                'accepted'      :{'type':'integer'},
                'errorEvents'   :{'type':'integer'},
-               'size'          :{'type':'integer'},
+               'size'          :{'type':'long'},
             }
          }
       }
@@ -106,7 +106,7 @@ def esMonitorMapping(esServerUrl,esIndexName,numberOfShards,numberOfReplicas,deb
                'processed'     :{'type':'integer'},
                'accepted'      :{'type':'integer'},
                'errorEvents'   :{'type':'integer'},
-               'size'          :{'type':'integer'},
+               'size'          :{'type':'long'},
             }
          }
       }
